Separate the local storage of tasks
Created storageName variable that will store the name of tasks storage;
This will be then be used to separate the storage of the task module and
every project created.
Created at setStorageName function to be able to set the storageName.
Added a logic in the project Module to call on setStorageName function
to immediately create a different task storage for the project.
Revised each localStorage methods to call on the storageName.