Create storage for projects
Created a baseProjectList variable to store base properties of projects.
Created an updatedProjectsList variable to store projects with methods.
Created functions that will store/get baseProjectList to local storage;
Namely addProjectToStorage, getProjectFromStorage, updateProjectStorage.
Divided the project factory functions into 2, one for creating a base
Project, the other is to apply methods.
Created applyProjectMethods to apply methods to each project of
baseProjectList.
Revised the listProjects function to call on the applyProjectMethods and
return the updatedProjectsList.
Revised the deleteProject function to delete the project from both
updatedProjectsList and baseProjectList.
Created a function to loadStorage and loadProjectStorage to use local
storage upon page load.